requirements.txt
Flask==3.0.0
nltk==3.8.1
requests==2.31.0
scikit-learn==1.4.0
spacy==3.7.2
python-dotenv==1.0.0
newsapi-python==0.2.6
wikipedia==1.4.0